OS based aliases!

+++ .zshrc
@@ -1,32 +1,43 @@
-
-# If you come from bash you might have to change your $PATH.
+# $PATH variable
 export PATH=/opt/homebrew/opt/ruby/bin:$HOME/bin:/usr/local/bin:$HOME/.cargo/bin:$PATH
 
-# Path to your oh-my-zsh installation.
+# zsh path
 export ZSH="$HOME/.oh-my-zsh"
 
-# Set name of the theme to load --- if set to "random", it will
-# load a random theme each time oh-my-zsh is loaded, in which case,
-# to know which specific one was loaded, run: echo $RANDOM_THEME
-# See https://github.com/ohmyzsh/ohmyzsh/wiki/Themes
+# neovim please
 export VISUAL=nvim
 export EDITOR=nvim
 
+# personalized directories
 export configd="$HOME/.config"
 export workd="$HOME/work"
 export gitd="$HOME/github"
 export smooth="$gitd/smooth-box"
 export tmuxd="$HOME/.config/tmux"
 
-# Aliases
-alias zedit='nvim $HOME/.zshrc'
+# generic linux detection
+if [[ "$OSTYPE" == *"linux-gnu"* ]]; then
+  echo 'linux-y'
+  alias cleansyslog='sudo truncate -s 0 /var/log/syslog'
+
+  if command -v sway > /dev/null; then
+    echo 'sway'
+    alias getids='swaymsg -t get_tree'
+    alias wayreset='killall -SIGUSR2 waybar &!'
+  fi
+fi
+
+# macOS detection
+if [[ $(uname) == "Darwin" ]]; then
+  echo 'Mac OS'
+fi
+
+# OS agnostic aliases
 alias zload='source $HOME/.zshrc'
+alias zedit='nvim $HOME/.zshrc'
 alias tedit='nvim $HOME/.tmux.conf'
 alias tkill='tmux kill-session -a'
 alias nedit='$HOME/.config/nvim/ && nvim .'
-alias cleansyslog='sudo truncate -s 0 /var/log/syslog'
-
-alias wayreset='killall -SIGUSR2 waybar &!'
 
 # check if tmux and launch
 if [[ -z "$TMUX" ]]; then
@@ -40,12 +51,8 @@ fi
 
 # alias fd="fdfind --type f --hidden --exclude .git | fzf-tmux -p --reverse | xargs nvim && clear"
 # alias ll="lsd -lah"
-#
-# Which plugins would you like to load?
-# Standard plugins can be found in $ZSH/plugins/
+
 # Custom plugins may be added to $ZSH_CUSTOM/plugins/
-# Example format: plugins=(rails git textmate ruby lighthouse)
-# Add wisely, as too many plugins slow down shell startup.
 plugins=(git fd fzf sudo archlinux zoxide)
 
 # powerlevel10k prompt
@@ -56,27 +63,22 @@ ZSH_THEME="powerlevel10k/powerlevel10k"
 [[ ! -f ~/.p10k.zsh ]] || source ~/.p10k.zsh
 
 source $ZSH/oh-my-zsh.sh
-# ZSH_THEME="spaceship"
 
 # bun
 export BUN_INSTALL="$HOME/.bun"
 export PATH="$BUN_INSTALL/bin:$PATH"
 
 # nvm
-# export NVM_DIR="$HOME/.nvm"
-# source /usr/share/nvm/init-nvm.sh
+export NVM_DIR="$HOME/.nvm"
+source /usr/share/nvm/init-nvm.sh
 
 
 # NOTE: macOS nvm configs via homebrew
-[ -s "/opt/homebrew/opt/nvm/nvm.sh" ] && \. "/opt/homebrew/opt/nvm/nvm.sh"  # This loads nvm
-[ -s "/opt/homebrew/opt/nvm/etc/bash_completion.d/nvm" ] && \. "/opt/homebrew/opt/nvm/etc/bash_completion.d/nvm"  # This loads nvm bash_completion
+# [ -s "/opt/homebrew/opt/nvm/nvm.sh" ] && \. "/opt/homebrew/opt/nvm/nvm.sh"  # This loads nvm
+# [ -s "/opt/homebrew/opt/nvm/etc/bash_completion.d/nvm" ] && \. "/opt/homebrew/opt/nvm/etc/bash_completion.d/nvm"  # This loads nvm bash_completion
 
 # WARN: Figure out an actual ruby solution? why is it so hard to get this dang language usable
-# source $HOME/.rvm/scripts/rvm
-
-# supposed flags for ruby
-# export LDFLAGS="-L/opt/homebrew/opt/ruby/lib"
-# export CPPFLAGS="-I/opt/homebrew/opt/ruby/include"
+source $HOME/.rvm/scripts/rvm
 
 # bun completions
 [ -s "/home/justinprime/.bun/_bun" ] && source "/home/justinprime/.bun/_bun"
